+++ src/components/RowMenu/index.tsx
@@ -11,7 +11,7 @@ import {
 	openMenu,
 	closeTopLevelMenu,
 	isMenuOpen,
-} from "src/services/redux/menuSlice";
+} from "src/services/menu/menuSlice";
 import { numToPx, pxToNum } from "src/services/string/conversion";
 
 import "./styles.css";
